로그인 & AI 구현 & 영화추천 완료

+++ backend/reviews/views.py
@@ -8,27 +8,32 @@ from django.shortcuts import get_object_or_404
 
 class ReviewListCreateAPIView(generics.ListCreateAPIView):
     serializer_class = ReviewSerializer
-    permission_classes = [permissions.AllowAny]  # 🔥 전부 허용
+    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
 
     def get_queryset(self):
+        # Frontend passes TMDB ID as movie_id in URL
         movie_id = self.kwargs['movie_id']
-        return Review.objects.filter(movie_id=movie_id)
+        movie = get_object_or_404(Movie, tmdb_id=movie_id)
+        return Review.objects.filter(movie=movie)
 
     def perform_create(self, serializer):
-        movie = get_object_or_404(Movie, id=self.kwargs['movie_id'])
-        serializer.save(movie=movie)  # 🔥 user 안 넣음
+        movie_id = self.kwargs['movie_id']
+        movie = get_object_or_404(Movie, tmdb_id=movie_id)
+        serializer.save(user=self.request.user, movie=movie)
 
 class ReviewDetailAPIView(generics.RetrieveUpdateDestroyAPIView):
     serializer_class = ReviewSerializer
-    permission_classes = [permissions.AllowAny]  # 🔥 전부 허용
+    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
 
     def get_queryset(self):
         return Review.objects.all()
 
-    # 🔥 수정 제한 없음
     def perform_update(self, serializer):
+        if self.get_object().user != self.request.user:
+            raise PermissionDenied("수정 권한이 없습니다.")
         serializer.save()
 
-    # 🔥 삭제 제한 없음
     def perform_destroy(self, instance):
+        if instance.user != self.request.user:
+            raise PermissionDenied("삭제 권한이 없습니다.")
         instance.delete()
